backend api connected and header page coded

+++ backend_files/cash_candy_app/models.py
@@ -41,7 +41,7 @@ class AllowanceDetail(models.Model):
     transaction_description = models.TextField(null=False)
     allowancelist = models.ForeignKey(AllowanceList, related_name="allowance_detail", on_delete=models.CASCADE)  #one list to many transactions - foreign key
     transaction_date = models.DateTimeField(null=False)
-    transaction_amount = models.IntegerField(null=False)    
+    transaction_amount = models.DecimalField(max_digits=6, decimal_places=2, null=False)    
 
     def __str__(self):
         return f"{self.transaction_date, self.transaction_description}"
